pip-requirements.txt
appdirs==1.4.0
autopep8==1.3.2
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.3
crypto==1.4.1
Django==1.8
django-crispy-forms==1.6.1
django-filter==1.0.2
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-registration-redux==1.4
djangorestframework==3.6.3
et-xmlfile==1.0.1
idna==2.5
jdcal==1.3
Markdown==2.6.8
markdown2==2.3.4
mobmsg==0.1
MySQL-python==1.2.5
Naked==0.1.31
numpy==1.13.0
olefile==0.44
openpyxl==2.4.8
packaging==16.8
pandas==0.20.2
pbr==3.0.1
Pillow==4.1.1
pkginfo==1.4.1
pycodestyle==2.3.1
pycrypto==2.6.1
PyJWT==1.5.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.14.1
requests-toolbelt==0.8.0
shellescape==3.4.1
six==1.10.0
stevedore==1.21.0
tqdm==4.14.0
twine==1.9.1
urllib3==1.21.1
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
xlrd==1.0.0
